Add Chassis::Serializable
This is a general purpose module that implements serialization
using `marshal_dump` and `marshal_load`. JSON and YAML delegate
to this interface after loading the Hash.

This module's purpose is to allow objects to be serialized across
multiple formats to work with multiple date stores like
Memcached, NoSQL, PStore, and even PostgreSQL with JSON column
or standard serialization approaches. This should be useful in
future persistence implementations.