Fix ParameterMissingError message
It referred to an instance variable instead of the block variable
causing this error to return a very confusing message.

I have looked through the tests and see there are no specific tests for
the error messages. Should it be added?

Perhaps just something that checks that the key is contained within the
error message.